stencil/src/components/my-counter/my-counter.spec.ts
import { MyCounter } from './my-counter';

describe('my-counter', () => {
  it('builds', () => {
    expect(new MyCounter()).toBeTruthy();
  });

  describe('formatting', () => {
    it('returns empty string for no names defined', () => {
      const component = new MyCounter();
      expect(component.format()).toEqual('');
    });

    it('formats just first names', () => {
      const component = new MyCounter();
      component.first = 'Joseph';
      expect(component.format()).toEqual('Joseph');
    });

    it('formats first and last names', () => {
      const component = new MyCounter();
      component.first = 'Joseph';
      component.last = 'Publique';
      expect(component.format()).toEqual('Joseph Publique');
    });

    it('formats first, middle and last names', () => {
      const component = new MyCounter();
      component.first = 'Joseph';
      component.middle = 'Quincy';
      component.last = 'Publique';
      expect(component.format()).toEqual('Joseph Quincy Publique');
    });
  });
});
